TipsterArena/requirements.txt
autopep8==1.4.4
click==8.1.7
colorama==0.4.6
entrypoints==0.3
exceptiongroup==1.1.3
flake8==3.7.7
Flask==2.2.5
Flask-Login==0.6.2
flask-talisman==1.1.0
Flask-WTF==1.1.1
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==3.0.0
importlib-metadata==6.7.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.6.1
packaging==23.2
pluggy==1.2.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==0.21.1
tomli==2.0.1
typing_extensions==4.7.1
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
zope.event==5.0
zope.interface==6.1
